requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
click==8.1.8
distro==1.9.0
Django==5.1
djangorestframework==3.16.0
exceptiongroup==1.2.2
Faker==37.1.0
groq==0.23.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.7
pydantic==2.11.3
pydantic_core==2.33.1
sniffio==1.3.1
sqlparse==0.5.1
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
